+++ aws_cfn_ses_domain/lambda_function.py
@@ -21,11 +21,8 @@ DEFAULT_PROPERTIES = {
     "MailFromSubdomain": "mail",
     "CustomDMARC": '"v=DMARC1; p=none; pct=100; sp=none; aspf=r;"',
     "TTL": "1800",
-    "Region": os.getenv("AWS_REGION"),
+    "Region": os.getenv("AWS_REGION"),  # where the stack (lambda fn) is running
 }
-
-
-ses = boto3.client('ses')
 
 
 def lambda_handler(event, context):
@@ -47,6 +44,23 @@ def lambda_handler(event, context):
                     reason="The 'Domain' property is required.",
                     physical_resource_id="MISSING")
 
+    # Use an SES Identity ARN as the PhysicalResourceId - see:
+    # https://docs.aws.amazon.com/IAM/latest/UserGuide/list_amazonses.html#amazonses-resources-for-iam-policies
+    domain_arn = format_arn(
+        service="ses", region=properties["Region"],
+        resource_type="identity", resource_name=domain,
+        defaults_from=event["StackId"])  # current stack's ARN has account and partition
+
+    if event["RequestType"] == "Delete" and event["PhysicalResourceId"] == domain:
+        # v0.3 backwards compatibility:
+        # Earlier versions used just the domain as the PhysicalResourceId.
+        # When a CF update results in a new v0.3 id (ARN, rather than domain), CF will
+        # automatically issue a Delete on the old id. We need to ignore that
+        # request (or we'd incorrectly delete the domain we meant to provision).
+        return send(event, context, SUCCESS,
+                    response_data={"Domain": domain},
+                    physical_resource_id=domain)
+
     if event["RequestType"] == "Delete":
         # Treat Delete as a request to disable both directions
         properties["EnableSend"] = False
@@ -59,22 +73,26 @@ def lambda_handler(event, context):
         # for ClientError, might be helpful to look at error.response, too
         logger.exception("Error updating SES: %s", error)
         return send(event, context, FAILED,
-                    reason=str(error), physical_resource_id=domain)
+                    reason=str(error), physical_resource_id=domain_arn)
 
     # Determine required DNS
     properties.update(outputs)
     route53_records = generate_route53_records(properties)
     outputs.update({
+        "Arn": domain_arn,
         "Domain": domain,
+        "Region": properties["Region"],
         "Route53RecordSets": route53_records,
         "ZoneFileEntries": route53_to_zone_file(route53_records),
     })
 
     return send(event, context, SUCCESS,
-                response_data=outputs, physical_resource_id=domain)
+                response_data=outputs, physical_resource_id=domain_arn)
 
 
 def update_ses_domain_identity(domain, properties):
+    ses = boto3.client('ses', region_name=properties['Region'])
+
     """Handle SES (de-)provisioning for domain and returns dict of output info"""
     outputs = {}
     enable_send = properties["EnableSend"]
@@ -177,3 +195,32 @@ def route53_to_zone_file(records):
             ttl=record["TTL"], type=record["Type"],
             data=" ".join(record["ResourceRecords"]))
         for record in records]
+
+
+def format_arn(partition=None, service=None, region=None, account=None,
+               resource=None, resource_type=None, resource_name=None,
+               defaults_from=None):
+    """Return an ARN composed from the specified components.
+
+    Provide either resource or both resource_type and resource_name.
+
+    defaults_from can be an existing ARN, which will be used to fill in any
+    missing components.
+
+    See https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_identifiers.html
+    for the format.
+    """
+    if resource is None and resource_type is not None:
+        resource = f"{resource_type}/{resource_name}"
+    if defaults_from is not None:
+        try:
+            _arn, _partition, _service, _region, _account, _resource = defaults_from.split(":")
+        except (TypeError, ValueError):
+            raise ValueError(f"Invalid ARN in defaults_from={defaults_from!r}")
+        partition = partition if partition is not None else _partition
+        service = service if service is not None else _service
+        region = region if region is not None else _region
+        account = account if account is not None else _account
+        resource = resource if resource is not None else _resource
+
+    return f"arn:{partition}:{service}:{region}:{account}:{resource}"