+++ tests/test_lambda_handler.py
@@ -1,13 +1,15 @@
 import os
+import boto3
 from unittest import TestCase
 from unittest.mock import patch, ANY as MOCK_ANY
 from botocore.stub import Stubber
 
 os.environ["AWS_REGION"] = "mock-region"  # (before loading lambda_function)
-from aws_cfn_ses_domain.lambda_function import lambda_handler, ses
+from aws_cfn_ses_domain.lambda_function import lambda_handler
 
 
 mock_context = object()
+mock_stack_id = "arn:aws:cloudformation:mock-region:111111111111:stack/example/deadbeef"
 
 
 class TestLambdaHandler(TestCase):
@@ -16,9 +18,15 @@ class TestLambdaHandler(TestCase):
     maxDiff = None
 
     def setUp(self):
+        ses = boto3.client('ses')  # need a real client for Stubber
+        boto3_client_patcher = patch('aws_cfn_ses_domain.lambda_function.boto3.client', return_value=ses)
+        self.mock_boto3_client = boto3_client_patcher.start()
+        self.addCleanup(boto3_client_patcher.stop)
+
         self.ses_stubber = Stubber(ses)
         self.ses_stubber.activate()
         self.addCleanup(self.ses_stubber.deactivate)
+
         send_patcher = patch('aws_cfn_ses_domain.lambda_function.send')
         self.mock_send = send_patcher.start()
         self.addCleanup(send_patcher.stop)
@@ -36,7 +44,8 @@ class TestLambdaHandler(TestCase):
     def test_domain_required(self):
         event = {
             "RequestType": "Create",
-            "ResourceProperties": {}}
+            "ResourceProperties": {},
+            "StackId": mock_stack_id}
         lambda_handler(event, mock_context)
         self.assertLambdaResponse(
             event, status="FAILED",
@@ -48,7 +57,8 @@ class TestLambdaHandler(TestCase):
             "RequestType": "Create",
             "ResourceProperties": {
                 "Domain": " . ",
-            }}
+            },
+            "StackId": mock_stack_id}
         lambda_handler(event, mock_context)
         self.assertLambdaResponse(
             event, status="FAILED",
@@ -60,7 +70,8 @@ class TestLambdaHandler(TestCase):
             "RequestType": "Create",
             "ResourceProperties": {
                 "Domain": "example.com.",
-            }}
+            },
+            "StackId": mock_stack_id}
         self.ses_stubber.add_response(
             'verify_domain_identity',
             {'VerificationToken': "ID_TOKEN"},
@@ -75,7 +86,11 @@ class TestLambdaHandler(TestCase):
             {'Identity': "example.com", 'MailFromDomain': "mail.example.com"})
         lambda_handler(event, mock_context)
 
-        outputs = self.assertLambdaResponse(event, physical_resource_id="example.com")
+        # Should default to SES in current region (where stack is running):
+        self.mock_boto3_client.assert_called_once_with('ses', region_name="mock-region")
+
+        outputs = self.assertLambdaResponse(
+            event, physical_resource_id="arn:aws:ses:mock-region:111111111111:identity/example.com")
         self.assertEqual(outputs["Domain"], "example.com")
         self.assertEqual(outputs["VerificationToken"], "ID_TOKEN")
         self.assertEqual(outputs["DkimTokens"], ["DKIM_TOKEN_1", "DKIM_TOKEN_2"])
@@ -106,6 +121,8 @@ class TestLambdaHandler(TestCase):
             'mail.example.com.                   \t1800\tIN\tTXT  \t"v=spf1 include:amazonses.com -all"',
             '_dmarc.example.com.                 \t1800\tIN\tTXT  \t"v=DMARC1; p=none; pct=100; sp=none; aspf=r;"',
         ])
+        self.assertEqual(outputs["Region"], "mock-region")
+        self.assertEqual(outputs["Arn"], "arn:aws:ses:mock-region:111111111111:identity/example.com")
 
     def test_create_all_options(self):
         event = {
@@ -118,7 +135,8 @@ class TestLambdaHandler(TestCase):
                 "CustomDMARC": '"v=DMARC1; p=quarantine; rua=mailto:d@example.com;"',
                 "TTL": "300",
                 "Region": "us-test-2",
-            }}
+            },
+            "StackId": mock_stack_id}
         self.ses_stubber.add_response(
             'verify_domain_identity',
             {'VerificationToken': "ID_TOKEN"},
@@ -133,7 +151,11 @@ class TestLambdaHandler(TestCase):
             {'Identity': "example.com", 'MailFromDomain': "bounce.example.com"})
         lambda_handler(event, mock_context)
 
-        outputs = self.assertLambdaResponse(event, physical_resource_id="example.com")
+        # Should override SES region when Region property provided:
+        self.mock_boto3_client.assert_called_once_with('ses', region_name="us-test-2")
+
+        outputs = self.assertLambdaResponse(
+            event, physical_resource_id="arn:aws:ses:us-test-2:111111111111:identity/example.com")
         self.assertEqual(outputs["Domain"], "example.com")
         self.assertEqual(outputs["VerificationToken"], "ID_TOKEN")
         self.assertEqual(outputs["DkimTokens"], ["DKIM_TOKEN_1", "DKIM_TOKEN_2"])
@@ -167,16 +189,19 @@ class TestLambdaHandler(TestCase):
             '_dmarc.example.com.                 \t300\tIN\tTXT  \t"v=DMARC1; p=quarantine; rua=mailto:d@example.com;"',
             'example.com.                        \t300\tIN\tMX   \t10 inbound-smtp.us-test-2.amazonaws.com.',
         ])
+        self.assertEqual(outputs["Region"], "us-test-2")
+        self.assertEqual(outputs["Arn"], "arn:aws:ses:us-test-2:111111111111:identity/example.com")
 
     def test_update_receive_only(self):
         event = {
-            "RequestType": "Create",
+            "RequestType": "Update",
             "ResourceProperties": {
                 "Domain": "example.com.",
                 "EnableSend": False,
                 "EnableReceive": True,
                 "CustomDMARC": None,
-            }}
+            },
+            "StackId": mock_stack_id}
         self.ses_stubber.add_response(
             'verify_domain_identity',
             {'VerificationToken': "ID_TOKEN"},
@@ -187,7 +212,8 @@ class TestLambdaHandler(TestCase):
             {'Identity': "example.com", 'MailFromDomain': ""})
         lambda_handler(event, mock_context)
 
-        outputs = self.assertLambdaResponse(event, physical_resource_id="example.com")
+        outputs = self.assertLambdaResponse(
+            event, physical_resource_id="arn:aws:ses:mock-region:111111111111:identity/example.com")
         self.assertEqual(outputs["Domain"], "example.com")
         self.assertEqual(outputs["VerificationToken"], "ID_TOKEN")
         self.assertNotIn("DkimTokens", outputs)
@@ -210,11 +236,13 @@ class TestLambdaHandler(TestCase):
     def test_delete(self):
         event = {
             "RequestType": "Delete",
+            "PhysicalResourceId": "arn:aws:ses:mock-region:111111111111:identity/example.com",
             "ResourceProperties": {
                 "Domain": "example.com.",
                 "EnableSend": True,
                 "EnableReceive": True,
-            }}
+            },
+            "StackId": mock_stack_id}
         self.ses_stubber.add_response(
             'delete_identity',
             {},
@@ -225,7 +253,8 @@ class TestLambdaHandler(TestCase):
             {'Identity': "example.com", 'MailFromDomain': ""})
         lambda_handler(event, mock_context)
 
-        outputs = self.assertLambdaResponse(event, physical_resource_id="example.com")
+        outputs = self.assertLambdaResponse(
+            event, physical_resource_id="arn:aws:ses:mock-region:111111111111:identity/example.com")
         self.assertEqual(outputs["Domain"], "example.com")
         self.assertNotIn("VerificationToken", outputs)
         self.assertNotIn("DkimTokens", outputs)
@@ -237,12 +266,32 @@ class TestLambdaHandler(TestCase):
         self.assertEqual(outputs["Route53RecordSets"], [])
         self.assertEqual(outputs["ZoneFileEntries"], [])
 
+    def test_v0_3_physical_id_change(self):
+        # Prior to v0.3, the PhysicalResourceId was just the cleaned Domain.
+        # Make sure we ignore the Delete operation that CF will issue on the
+        # old physical ID after upgrading.
+        event = {
+            "RequestType": "Delete",
+            "PhysicalResourceId": "example.com",  # old format: just the domain
+            "ResourceProperties": {
+                "Domain": "example.com.",
+                "EnableSend": True,
+                "EnableReceive": True,
+            },
+            "StackId": mock_stack_id}
+        # self.ses_stubber.nothing: *no* SES ops should occur
+        lambda_handler(event, mock_context)
+
+        outputs = self.assertLambdaResponse(event, physical_resource_id="example.com")
+        self.assertEqual(outputs["Domain"], "example.com")
+
     def test_boto_error(self):
         event = {
             "RequestType": "Create",
             "ResourceProperties": {
                 "Domain": "bad domain name",
-            }}
+            },
+            "StackId": mock_stack_id}
         self.ses_stubber.add_client_error(
             'verify_domain_identity',
             "InvalidParameterValue",
@@ -254,7 +303,7 @@ class TestLambdaHandler(TestCase):
             event, status="FAILED",
             reason="An error occurred (InvalidParameterValue) when calling the"
                    " VerifyDomainIdentity operation: Invalid domain name bad domain name.",
-            physical_resource_id="bad domain name")
+            physical_resource_id=MOCK_ANY)
 
         # Check that the exception got logged
         self.assertEqual(len(cm.output), 1)