+++ CHANGELOG.md
@@ -2,7 +2,30 @@
 
 ## v0.3.dev0
 
-*(Development testing)*
+*Unreleased changes*
+
+### Breaking changes
+
+* The physical resource ID of a `Custom::SES_Domain` resource is now the full ARN 
+  of the Amazon SES domain identity. (Previously it was just the domain name.)
+  If you have templates that use `!Ref MySESDomain` to get the domain, change them 
+  to `!GetAtt MySESDomain.Domain` instead.
+
+* Because of the physical resource ID change, the first stack update after upgrading
+  to v0.3 will appear to update and then delete existing `Custom::SES_Domain` resources.
+  Don't be alarmed: your SES domain *isn't* actually deleted. (The delete operation 
+  covers the old physical resource ID, and will be ignored.)
+
+### Fixes
+
+* Support using the [`Region`](README.md#region) property to provision an Amazon SES 
+  domain in a different region from where you're running your CloudFormation stack.
+  (Thanks to @gfodor.)
+
+### Features
+
+* Make [`Arn`](README.md#other-attributes) and [`Region`](README.md#other-attributes)
+  attributes available on `Custom::SES_Domain` resources.
 
 
 ## v0.2